feat: Sessions 2-13 — full-stack MVP + Swarm Engine
Frontend (apps/web/):
- 20 pages: landing, auth, dashboard, profile, assessment, events, leaderboard, notifications, settings, organizations
- 38+ components with Stitch design system (dark Material 3)
- i18n: AZ + EN with full translation coverage
- Supabase auth integration, TanStack Query, Zustand stores

Backend (apps/api/):
- 35 API endpoints: auth, profiles, assessment, AURA, events, organizations, badges, verification
- IRT/CAT adaptive assessment engine (3PL + EAP, pure Python)
- AURA score calculation with 8 weighted competencies
- Anti-gaming detection, reliability scoring
- Rate limiting, security headers, JWT verification via service role

Database (supabase/):
- 13 migrations: profiles, competencies, questions, sessions, AURA scores, badges, organizations, events, registrations, behavior signals, embeddings, RPC functions, expert verifications
- Seed data: 8 competencies, 6 badges, 20 questions (5 calibrated + 15 placeholder)

Swarm Engine (packages/swarm/):
- MiroFish Swarm Decision Engine v2 (~3100 lines)
- 15 providers auto-discovered from 3 API keys
- Skills library, agent memory, innovation field
- Conditional debate, LLM synthesis, self-learning calibration

+++ supabase/migrations/20260321000008_create_organizations.sql
@@ -40,7 +40,7 @@ SELECT
     organization_id,
     COUNT(*) AS rating_count,
     CASE WHEN COUNT(*) >= 5
-        THEN ROUND(AVG(rating) * 20, 1)  -- 1-5 → 0-100
+        THEN ROUND((AVG(rating) * 20)::numeric, 1)  -- 1-5 → 0-100
         ELSE NULL                          -- hidden until 5+ ratings
     END AS trust_score
 FROM public.organization_ratings
